Refactor builtin into smaller modules
